Fix call graph bug, improve filtering and add cursor printing
Most of calls were not graphed because of an incorrect reliance on
reference expressions.

The --src-file filter option facilitates selecting a specific file to
parse.  Also allow for the combination of filter options so that
header directories along with source directories can be selected

Add the ability to print cursors as they're encountered to help with
debug.